Simplify the implementation
The new architecture simply hooks `pre-redisplay-function` but uses a *snapshot*
of the windows of the current frame as a guard to avoid useless relayouts.

In doing so, the `track-mouse` issue (see #14 and #15) has been fixed for Emacs
version 25.

The rationale behind this has been discussed in the [emacs-devel][1] mailing
list.

[1]: http://lists.gnu.org/archive/html/emacs-devel/2018-05/msg00026.html