Introduce 'zoom-use-total-window-size'
This controls whether margins, header-line, etc. are used when computing the
window sizes. Before, that was the case only when floats were used to specify
the ratio. It makes sense instead to let the user decide, and also to change the
default to always include margins and other decorations in the window size
computation.

Close #42, as it provides a better implementation.